// src/app/core/enums/genric.enum.ts
export enum LanguageEnum {
    English = 'en',
    Germany = 'gr',
    Arabic = 'ar'
}

export enum GenderEnum {
    Male = 1,
    Female = 2
}

export enum PopupTypeEnum {
    Success = 1,
    Error = 2,
    Exclamation = 3
}

export enum PageEditingTypeEnum {
    Add = 1,
    Edit = 2,
    View = 3
}

export enum DefinitionFilesTypesEnum {
    ContractType = 1,
    IncomeType = 2,
    OutgoingType = 3
}

export enum UserActionsEnum {
    Post,
    Activity
}

export enum PagesEnum {
    RoleManagment = 'RoleManagment',
    UserManagment = 'UserManagment',
    DefinitionFiles = 'DefinitionFiles'
}